feat(Album View Page): Next/Prev arrows will load cached photo image when possible
