+++ ui/app/containers/App/reducer.js
@@ -11,8 +11,6 @@ import {
   LOAD_NEXT_THUMB_PAGE_SUCCESS,
   LOAD_THUMBS_SUCCESS,
   PAGE_SIZE,
-  NEXT_MEMORY,
-  PREV_MEMORY,
 } from '../AlbumViewPage/constants';
 import { insertPage } from '../AlbumViewPage/paging';
 import { LOAD_GALLERY, LOAD_GALLERY_SUCCESS } from '../GalleryViewPage/constants';
@@ -101,26 +99,6 @@ const appReducer = (state = initialState, action) => produce(state, (draft) => {
       draft[action.host][action.gallery][action.album].memories[memoryIndex].photoLink = action.photoLink;
 
       draft.currentMemory.photoLink = action.photoLink;
-      break;
-    }
-
-    case NEXT_MEMORY:
-    case PREV_MEMORY: {
-      const { memories } = state[state.host][state.gallery][state.album];
-      const currentMemoryId = state.currentMemory.id || 0;
-      const currentMemoryIndex = memories.findIndex(item => item.id === currentMemoryId);
-      let adjacentMemoryIndex = currentMemoryIndex + action.adjacentInt;
-
-      const carouselEnd = adjacentMemoryIndex >= memories.length;
-      if (carouselEnd) adjacentMemoryIndex -= memories.length;
-
-      const carouselBegin = adjacentMemoryIndex < 0;
-      if (carouselBegin) adjacentMemoryIndex = memories.length + adjacentMemoryIndex;
-
-      const findIndex = memories[adjacentMemoryIndex].id;
-
-      const found = state[state.host][state.gallery][state.album].memories.filter(item => item.id === findIndex)[0];
-      draft.currentMemory = found || {};
       break;
     }
